requirements.txt
rioxarray
xarray
dask
numpy
pandas
geopandas
requests
folium
flask
flask-mail
flask-wtf
flask-cors
gunicorn
werkzeug
email-validator
netCDF4
scipy
google-cloud-storage
area
#requests-toolbelt==0.10.0
#Pyrebase4
#firebase-admin
matplotlib